+++ message/domains/__init__.py
@@ -1,0 +1,2 @@
+from .message_models import Message, MessageRoom
+from .message_block_models import MessageBlock
