+++ message/views/__init__.py
@@ -1,0 +1,2 @@
+from .message_CRUD_views import *
+from .message_block_views import *
